Implements a database to store the notes

+++ app/src/main/java/com/huggets/mynotes/note/NoteItemUiState.kt
@@ -1,14 +1,14 @@
 package com.huggets.mynotes.note
 
 data class NoteItemUiState(
-    val id: Int,
+    val id: Long,
     val title: String,
     val content: String
 ) {
     fun toNote() = Note(id, title, content)
 }
 
-fun List<NoteItemUiState>.find(noteId: Int): NoteItemUiState? {
+fun List<NoteItemUiState>.find(noteId: Long): NoteItemUiState? {
     forEach {
         if (it.id == noteId) return it
     }
